+++ parser_commons/src/main/java/org/eclipse/moquette/proto/messages/MessageIDMessage.java
@@ -24,13 +24,13 @@ package org.eclipse.moquette.proto.messages;
  * @author andrea
  */
 public abstract class MessageIDMessage extends AbstractMessage {
-    private Integer m_messageID; //could be null if Qos is == 0
+    private Long m_messageID; //could be null if Qos is == 0
 
-    public Integer getMessageID() {
+    public Long getMessageID() {
         return m_messageID;
     }
 
-    public void setMessageID(Integer messageID) {
+    public void setMessageID(Long messageID) {
         this.m_messageID = messageID;
     }
